rock paper scissors - updates
